Estimates the significance of the observed Kullback-Leibler divergence by comparig to randomizations.

Estimates the significance of the observed Kullback-Leibler divergence by comparig to randomizations.

get_log_p_D_KL(T.counts, D_KL.observed, D_KL.randomized, output.dir = NULL)

Arguments

T.counts
The number of cells in which a gene is detected.

D_KL.observed
A vector of observed Kullback-Leibler divergences.

D_KL.randomized
A matrix of Kullback-Leibler divergences of randomized datasets.

output.dir
Optional parameter. Default is NULL. If not NULL, some files will be written to this directory.

Value

A vector of log10 p values, not corrected for multiple testing using the Bonferroni correction.