The main Haystack function.

The main Haystack function.

haystack(x, ...)

# S3 method for matrix
haystack(x, dim1 = 1, dim2 = 2, method = "highD",
  ...)

# S3 method for data.frame
haystack(x, ...)

# S3 method for Seurat
haystack(x, assay = "RNA", slot = "data",
  coord = "tsne", cutoff = 1, ...)

# S3 method for SingleCellExperiment
haystack(x, assay = "counts",
  coord = "TSNE", cutoff = 1, ...)

Arguments

x
a matrix or other object from which coordinates of cells can be extracted.

...
further paramters passed to haystack_raw().

dim1
column index or name of matrix for x-axis coordinates.

dim2
column index or name of matrix for y-axis coordinates.

method
choose between highD (default) and 2D haystack.

assay
name of assay data for Seurat method.

slot
name of slot for assay data for Seurat method.

coord
name of coordinates slot for specific methods.

cutoff
cutoff for detection.

Value

An object of class "haystack"