The main Haystack function, for 2-dimensional spaces.

The main Haystack function, for 2-dimensional spaces.

haystack_2D(
  x,
  y,
  detection,
  use.advanced.sampling = NULL,
  dir.randomization = NULL
)

Arguments

x
x-axis coordinates of cells in a 2D representation (e.g. resulting from PCA or t-SNE)

y
y-axis coordinates of cells in a 2D representation

detection
A logical matrix showing which genes (rows) are detected in which cells (columns)

use.advanced.sampling
If NULL naive sampling is used. If a vector is given (of length = no. of cells) sampling is done according to the values in the vector.

dir.randomization
If NULL, no output is made about the random sampling step. If not NULL, files related to the randomizations are printed to this directory.

Value

An object of class "haystack"

Examples

# using the toy example of the singleCellHaystack package
# define a logical matrix with detection of each gene (rows) in each cell (columns)
dat.detection <- dat.expression > 1

# running haystack in default mode
res <- haystack(dat.tsne, detection=dat.detection, method = "2D")

#> ### calling haystack_2D()...

#> ### setting parameters...

#> ### calculating Kullback-Leibler divergences...

#> ### starting randomizations...

#> ### ... 10 sets out of 58 done

#> ### ... 20 sets out of 58 done

#> ### ... 30 sets out of 58 done

#> ### ... 40 sets out of 58 done

#> ### ... 50 sets out of 58 done

#> ### estimating p-values...

#> ### returning result...

# list top 10 biased genes
show_result_haystack(res, n =10)

#>              D_KL log.p.vals log.p.adj T.counts
#> gene_242 1.360661  -26.25026 -23.55129       96
#> gene_275 1.345251  -25.57849 -22.87952       95
#> gene_317 1.386194  -25.49684 -22.79787       94
#> gene_244 1.238011  -24.45347 -21.75450       95
#> gene_497 1.599646  -24.10527 -21.40630       86
#> gene_351 1.408946  -23.57271 -20.87374       89
#> gene_137 1.349994  -23.03142 -20.33245       89
#> gene_339 1.414814  -22.90443 -20.20546       87
#> gene_429 1.071988  -22.12392 -19.42495       94
#> gene_79  1.929800  -22.05453 -19.35556       68