The main Haystack function, for higher-dimensional spaces and continuous expression levels.

The main Haystack function, for higher-dimensional spaces and continuous expression levels.

haystack_continuous_highD(
  x,
  expression,
  grid.points = 100,
  weights.advanced.Q = NULL,
  dir.randomization = NULL,
  scale = TRUE,
  grid.method = "centroid",
  randomization.count = 100,
  n.genes.to.randomize = 100,
  selection.method.genes.to.randomize = "heavytails",
  grid.coord = NULL,
  spline.method = "ns"
)

Arguments

x

Coordinates of cells in a 2D or higher-dimensional space. Rows represent cells, columns the dimensions of the space.

expression

a matrix with expression data of genes (rows) in cells (columns)

grid.points

An integer specifying the number of centers (grid points) to be used for estimating the density distributions of cells. Default is set to 100.

weights.advanced.Q

(Default: NULL) Optional weights of cells for calculating a weighted distribution of expression.

dir.randomization

If NULL, no output is made about the random sampling step. If not NULL, files related to the randomizations are printed to this directory.

scale

Logical (default=TRUE) indicating whether input coordinates in x should be scaled to mean 0 and standard deviation 1.

grid.method

The method to decide grid points for estimating the density in the high-dimensional space. Should be "centroid" (default) or "seeding".

randomization.count

Number of randomizations to use. Default: 100

n.genes.to.randomize

Number of genes to use in randomizations. Default: 100

selection.method.genes.to.randomize

Method used to select genes for randomization.

grid.coord

matrix of grid coordinates.

spline.method

Method to use for fitting splines "ns" (default): natural splines, "bs": B-splines.

Value

An object of class "haystack", including the results of the analysis, and the coordinates of the grid points used to estimate densities.

Examples

# using the toy example of the singleCellHaystack package

# running haystack
res <- haystack(dat.tsne, dat.expression)
#> ### calling haystack_continuous_highD()...
#> ### Using package sparseMatrixStats to speed up statistics in sparse matrices.
#> ### Calculating row-wise mean and SD... 
#> ### Filtered 0 genes with zero variance...
#> ### Using 100 randomizations...
#> ### Using 100 genes to randomize...
#> Warning: The value of 'grid.points' appears to be very high (> No. of cells / 10). You can set the number of grid points using the 'grid.points' parameter.
#> ### scaling input data...
#> ### deciding grid points...
#> ### calculating Kullback-Leibler divergences...
#> ### performing randomizations...
#> ### estimating p-values...
#> ### picking model for mean D_KL...
#> ### using natural splines
#> ### best RMSD  : 0.087
#> ### best df    : 3
#> ### picking model for stdev D_KL...
#> ### using natural splines
#> ### best RMSD  : 0.018
#> ### best df    : 4
#> ### returning result...
# list top 10 biased genes
show_result_haystack(res, n=10)
#>              D_KL log.p.vals log.p.adj
#> gene_242 1.931842  -38.38206 -35.68309
#> gene_62  2.077420  -35.99526 -33.29629
#> gene_79  2.236998  -35.05571 -32.35674
#> gene_444 2.031087  -34.16900 -31.47003
#> gene_275 1.833076  -33.81292 -31.11395
#> gene_61  2.009848  -33.43691 -30.73794
#> gene_351 1.889020  -33.41491 -30.71594
#> gene_497 1.867777  -31.80310 -29.10413
#> gene_99  1.970818  -31.73669 -29.03772
#> gene_429 1.748182  -31.68735 -28.98838