The main Haystack function, for higher-dimensional spaces.

The main Haystack function, for higher-dimensional spaces.

haystack_highD(
  x,
  detection,
  grid.points = 100,
  use.advanced.sampling = NULL,
  dir.randomization = NULL,
  scale = TRUE,
  grid.method = "centroid"
)

Arguments

x

Coordinates of cells in a 2D or higher-dimensional space. Rows represent cells, columns the dimensions of the space.

detection

A logical matrix showing which genes (rows) are detected in which cells (columns)

grid.points

An integer specifying the number of centers (grid points) to be used for estimating the density distributions of cells. Default is set to 100.

use.advanced.sampling

If NULL naive sampling is used. If a vector is given (of length = no. of cells) sampling is done according to the values in the vector.

dir.randomization

If NULL, no output is made about the random sampling step. If not NULL, files related to the randomizations are printed to this directory.

scale

Logical (default=TRUE) indicating whether input coordinates in x should be scaled to mean 0 and standard deviation 1.

grid.method

The method to decide grid points for estimating the density in the high-dimensional space. Should be "centroid" (default) or "seeding".

Value

An object of class "haystack", including the results of the analysis, and the coordinates of the grid points used to estimate densities.

Examples

# I need to add some examples.
# A toy example will be added too.