Function for hierarchical clustering of genes according to their distribution in a higher-dimensional space.

Function for hierarchical clustering of genes according to their distribution in a higher-dimensional space.

hclust_haystack_highD(
  x,
  detection,
  genes,
  method = "ward.D",
  grid.coordinates = NULL,
  scale = TRUE
)

Arguments

x
Coordinates of cells in a 2D or higher-dimensional space. Rows represent cells, columns the dimensions of the space.

detection
A logical matrix showing which genes (rows) are detected in which cells (columns)

genes
A set of genes (of the 'detection' data) which will be clustered.

method
The method to use for hierarchical clustering. See '?hclust' for more information. Default: "ward.D".

grid.coordinates
Coordinates of grid points in the same space as 'x', to be used to estimate densities for clustering.

scale
whether to scale data.

Value

An object of class hclust, describing a hierarchical clustering tree.

Examples

# to be added