Shows the results of the 'haystack' analysis in various ways, sorted by significance. Priority of params is genes > p.value.threshold > n.

Shows the results of the 'haystack' analysis in various ways, sorted by significance. Priority of params is genes > p.value.threshold > n.

show_result_haystack(res.haystack, n = NA, p.value.threshold = NA, gene = NA)

Arguments

res.haystack
A 'haystack' result variable

n
If defined, the top "n" sigificant genes will be returned. Default: NA, which shows all results.

p.value.threshold
If defined, genes passing this p-value threshold will be returned.

gene
If defined, the results of this (these) gene(s) will be returned.

Value

A table with a sorted subset of the 'haystack' result according to input parameters.

Examples

# using the toy example of the singleCellHaystack package
# define a logical matrix with detection of each gene (rows) in each cell (columns)
dat.detection <- dat.expression > 1

# running haystack in default mode
res <- haystack(dat.tsne, detection=dat.detection, method = "2D")

#> ### calling haystack_2D()...

#> ### setting parameters...

#> ### calculating Kullback-Leibler divergences...

#> ### starting randomizations...

#> ### ... 10 sets out of 58 done

#> ### ... 20 sets out of 58 done

#> ### ... 30 sets out of 58 done

#> ### ... 40 sets out of 58 done

#> ### ... 50 sets out of 58 done

#> ### estimating p-values...

#> ### returning result...

# below are variations for showing the results in a table
# 1. list top 10 biased genes
show_result_haystack(res.haystack = res, n =10)

#>              D_KL log.p.vals log.p.adj T.counts
#> gene_79  1.929800  -23.56713 -20.86816       68
#> gene_497 1.599646  -22.70672 -20.00775       86
#> gene_62  1.519993  -22.01876 -19.31979       81
#> gene_61  1.470475  -21.57234 -18.87337       80
#> gene_24  1.857497  -21.36552 -18.66655       59
#> gene_339 1.414814  -21.24188 -18.54291       87
#> gene_351 1.408946  -21.12828 -18.42931       89
#> gene_444 1.435882  -20.88703 -18.18806       75
#> gene_71  1.868822  -20.79010 -18.09113       56
#> gene_479 1.697257  -20.69243 -17.99346       61

# 2. list genes with p value below a certain threshold
show_result_haystack(res.haystack = res, p.value.threshold=1e-10)

#>               D_KL log.p.vals  log.p.adj T.counts
#> gene_79  1.9298000  -23.56713 -20.868161       68
#> gene_497 1.5996456  -22.70672 -20.007749       86
#> gene_62  1.5199933  -22.01876 -19.319789       81
#> gene_61  1.4704748  -21.57234 -18.873369       80
#> gene_24  1.8574968  -21.36552 -18.666549       59
#> gene_339 1.4148142  -21.24188 -18.542911       87
#> gene_351 1.4089458  -21.12828 -18.429312       89
#> gene_444 1.4358815  -20.88703 -18.188063       75
#> gene_71  1.8688220  -20.79010 -18.091131       56
#> gene_479 1.6972565  -20.69243 -17.993463       61
#> gene_137 1.3499940  -20.64356 -17.944589       89
#> gene_317 1.3861936  -20.58188 -17.882913       94
#> gene_300 1.6001648  -20.56509 -17.866124       64
#> gene_155 1.5347601  -20.25146 -17.552488       65
#> gene_242 1.3606607  -20.15957 -17.460601       96
#> gene_275 1.3452514  -20.15228 -17.453308       95
#> gene_381 1.6793541  -19.95565 -17.256683       58
#> gene_313 1.2945440  -19.94732 -17.248349       78
#> gene_57  1.8012672  -19.91044 -17.211474       54
#> gene_321 1.5354565  -19.89624 -17.197270       63
#> gene_59  1.4605859  -19.84402 -17.145046       66
#> gene_325 1.2093850  -19.46135 -16.762381       84
#> gene_244 1.2380108  -19.27029 -16.571320       95
#> gene_78  2.3302373  -19.08737 -16.388404       38
#> gene_463 2.1619665  -18.95536 -16.256389       41
#> gene_400 1.8579716  -18.89387 -16.194896       48
#> gene_112 1.4955762  -18.82462 -16.125651       59
#> gene_213 1.0901441  -18.33721 -15.638240       85
#> gene_135 1.7259726  -18.28989 -15.590919       49
#> gene_291 2.2015170  -17.94769 -15.248719       36
#> gene_340 1.4521210  -17.88765 -15.188675       56
#> gene_429 1.0719883  -17.87111 -15.172141       94
#> gene_488 3.0647537  -17.74930 -15.050334       22
#> gene_424 1.4472784  -17.43579 -14.736821       54
#> gene_99  1.0849556  -17.25026 -14.551293       71
#> gene_295 1.7966484  -17.11559 -14.416615       42
#> gene_458 0.9508191  -16.78738 -14.088412       92
#> gene_51  1.6407593  -16.36629 -13.667320       43
#> gene_449 2.4810190  -16.01835 -13.319383       24
#> gene_13  1.4534620  -14.44409 -11.745121       40
#> gene_10  1.3278710  -14.40565 -11.706682       44
#> gene_136 1.6923133  -13.69581 -10.996836       30
#> gene_216 1.7300714  -13.40599 -10.707021       28
#> gene_176 2.0046279  -13.10571 -10.406739       22
#> gene_156 1.4589040  -12.26021  -9.561241       30
#> gene_225 1.6076632  -11.29226  -8.593287       23
#> gene_201 1.2245458  -10.21066  -7.511689       28

# 3. list a set of specified genes
set <- c("gene_497","gene_386", "gene_275")
show_result_haystack(res.haystack = res, gene = set)

#>               D_KL log.p.vals log.p.adj T.counts
#> gene_497 1.5996456  -22.70672 -20.00775       86
#> gene_275 1.3452514  -20.15228 -17.45331       95
#> gene_386 0.2531526   -1.53752   0.00000       30