Shows the results of the 'haystack' analysis in various ways, sorted by significance. Priority of params is genes > p.value.threshold > n.

Shows the results of the 'haystack' analysis in various ways, sorted by significance. Priority of params is genes > p.value.threshold > n.

show_result_haystack(res.haystack, n = NA, p.value.threshold = NA, gene = NA)

Arguments

res.haystack
A 'haystack' result variable

n
If defined, the top "n" sigificant genes will be returned. Default: NA, which shows all results.

p.value.threshold
If defined, genes passing this p-value threshold will be returned.

gene
If defined, the results of this (these) gene(s) will be returned.

Value

A table with a sorted subset of the 'haystack' result according to input parameters.

Examples

# using the toy example of the singleCellHaystack package
# define a logical matrix with detection of each gene (rows) in each cell (columns)
dat.detection <- dat.expression > 1

# running haystack in default mode
res <- haystack(dat.tsne, detection=dat.detection, method = "2D")

#> ### calling haystack_2D()...

#> ### setting parameters...

#> ### calculating Kullback-Leibler divergences...

#> ### starting randomizations...

#> ### ... 10 values out of 58 done

#> ### ... 20 values out of 58 done

#> ### ... 30 values out of 58 done

#> ### ... 40 values out of 58 done

#> ### ... 50 values out of 58 done

#> ### estimating p-values...

#> ### returning result...

# below are variations for showing the results in a table
# 1. list top 10 biased genes
show_result_haystack(res.haystack = res, n =10)

#>              D_KL log.p.vals T.counts
#> gene_242 1.360661  -20.16854       96
#> gene_79  1.929800  -19.64425       68
#> gene_275 1.345251  -19.63955       95
#> gene_317 1.386194  -19.63668       94
#> gene_497 1.599646  -19.24491       86
#> gene_244 1.238011  -18.65685       95
#> gene_351 1.408946  -18.40704       89
#> gene_24  1.857497  -18.35954       59
#> gene_339 1.414814  -18.03468       87
#> gene_71  1.868822  -18.01217       56

# 2. list genes with p value below a certain threshold
show_result_haystack(res.haystack = res, p.value.threshold=1e-10)

#>               D_KL log.p.vals T.counts
#> gene_242 1.3606607  -20.16854       96
#> gene_79  1.9298000  -19.64425       68
#> gene_275 1.3452514  -19.63955       95
#> gene_317 1.3861936  -19.63668       94
#> gene_497 1.5996456  -19.24491       86
#> gene_244 1.2380108  -18.65685       95
#> gene_351 1.4089458  -18.40704       89
#> gene_24  1.8574968  -18.35954       59
#> gene_339 1.4148142  -18.03468       87
#> gene_71  1.8688220  -18.01217       56
#> gene_62  1.5199933  -17.92672       81
#> gene_137 1.3499940  -17.92402       89
#> gene_479 1.6972565  -17.53231       61
#> gene_61  1.4704748  -17.44594       80
#> gene_57  1.8012672  -17.25770       54
#> gene_300 1.6001648  -17.14474       64
#> gene_381 1.6793541  -17.04340       58
#> gene_78  2.3302373  -16.80313       38
#> gene_155 1.5347601  -16.75539       65
#> gene_444 1.4358815  -16.73419       75
#> gene_429 1.0719883  -16.68479       94
#> gene_463 2.1619665  -16.68100       41
#> gene_321 1.5354565  -16.58579       63
#> gene_400 1.8579716  -16.49781       48
#> gene_59  1.4605859  -16.28094       66
#> gene_313 1.2945440  -15.86586       78
#> gene_112 1.4955762  -15.85458       59
#> gene_135 1.7259726  -15.85141       49
#> gene_325 1.2093850  -15.82932       84
#> gene_291 2.2015170  -15.60750       36
#> gene_488 3.0647537  -15.36806       22
#> gene_340 1.4521210  -15.11944       56
#> gene_213 1.0901441  -14.89204       85
#> gene_458 0.9508191  -14.84208       92
#> gene_295 1.7966484  -14.79479       42
#> gene_424 1.4472784  -14.77613       54
#> gene_51  1.6407593  -14.02116       43
#> gene_449 2.4810190  -13.52906       24
#> gene_99  1.0849556  -13.49871       71
#> gene_13  1.4534620  -12.06403       40
#> gene_10  1.3278710  -12.00749       44
#> gene_136 1.6923133  -11.15721       30
#> gene_216 1.7300714  -10.84154       28
#> gene_176 2.0046279  -10.60677       22

# 3. list a set of specified genes
set <- c("gene_497","gene_386", "gene_275")
show_result_haystack(res.haystack = res, gene = set)

#>               D_KL log.p.vals T.counts
#> gene_275 1.3452514  -19.63955       95
#> gene_497 1.5996456  -19.24491       86
#> gene_386 0.2531526    0.00000       30