Shows the results of the 'haystack' analysis in various ways, sorted by significance. Priority of params is genes > p.value.threshold > n.

Shows the results of the 'haystack' analysis in various ways, sorted by significance. Priority of params is genes > p.value.threshold > n.

show_result_haystack(res.haystack, n = NA, p.value.threshold = NA, gene = NA)

Arguments

res.haystack
A 'haystack' result variable

n
If defined, the top "n" sigificant genes will be returned. Default: NA, which shows all results.

p.value.threshold
If defined, genes passing this p-value threshold will be returned.

gene
If defined, the results of this (these) gene(s) will be returned.

Value

A table with a sorted subset of the 'haystack' result according to input parameters.

Examples

# using the toy example of the singleCellHaystack package
# define a logical matrix with detection of each gene (rows) in each cell (columns)
dat.detection <- dat.expression > 1

# running haystack in default mode
res <- haystack(dat.tsne, detection=dat.detection, method = "2D")

#> ### calling haystack_2D()...

#> ### setting parameters...

#> ### calculating Kullback-Leibler divergences...

#> ### starting randomizations...

#> ### ... 10 values out of 58 done

#> ### ... 20 values out of 58 done

#> ### ... 30 values out of 58 done

#> ### ... 40 values out of 58 done

#> ### ... 50 values out of 58 done

#> ### estimating p-values...

#> ### returning result...

# below are variations for showing the results in a table
# 1. list top 10 biased genes
show_result_haystack(res.haystack = res, n =10)

#>              D_KL log.p.vals log.p.adj T.counts
#> gene_242 1.360661  -22.86751 -20.16854       96
#> gene_79  1.929800  -22.34322 -19.64425       68
#> gene_275 1.345251  -22.33852 -19.63955       95
#> gene_317 1.386194  -22.33565 -19.63668       94
#> gene_497 1.599646  -21.94388 -19.24491       86
#> gene_244 1.238011  -21.35582 -18.65685       95
#> gene_351 1.408946  -21.10601 -18.40704       89
#> gene_24  1.857497  -21.05851 -18.35954       59
#> gene_339 1.414814  -20.73365 -18.03468       87
#> gene_71  1.868822  -20.71114 -18.01217       56

# 2. list genes with p value below a certain threshold
show_result_haystack(res.haystack = res, p.value.threshold=1e-10)

#>               D_KL log.p.vals  log.p.adj T.counts
#> gene_242 1.3606607  -22.86751 -20.168540       96
#> gene_79  1.9298000  -22.34322 -19.644251       68
#> gene_275 1.3452514  -22.33852 -19.639546       95
#> gene_317 1.3861936  -22.33565 -19.636685       94
#> gene_497 1.5996456  -21.94388 -19.244915       86
#> gene_244 1.2380108  -21.35582 -18.656846       95
#> gene_351 1.4089458  -21.10601 -18.407036       89
#> gene_24  1.8574968  -21.05851 -18.359542       59
#> gene_339 1.4148142  -20.73365 -18.034676       87
#> gene_71  1.8688220  -20.71114 -18.012167       56
#> gene_62  1.5199933  -20.62569 -17.926725       81
#> gene_137 1.3499940  -20.62299 -17.924020       89
#> gene_479 1.6972565  -20.23128 -17.532310       61
#> gene_61  1.4704748  -20.14491 -17.445942       80
#> gene_57  1.8012672  -19.95667 -17.257699       54
#> gene_300 1.6001648  -19.84371 -17.144741       64
#> gene_381 1.6793541  -19.74237 -17.043396       58
#> gene_78  2.3302373  -19.50210 -16.803135       38
#> gene_155 1.5347601  -19.45436 -16.755387       65
#> gene_444 1.4358815  -19.43316 -16.734186       75
#> gene_429 1.0719883  -19.38376 -16.684795       94
#> gene_463 2.1619665  -19.37997 -16.680998       41
#> gene_321 1.5354565  -19.28476 -16.585788       63
#> gene_400 1.8579716  -19.19678 -16.497814       48
#> gene_59  1.4605859  -18.97991 -16.280937       66
#> gene_313 1.2945440  -18.56483 -15.865856       78
#> gene_112 1.4955762  -18.55355 -15.854584       59
#> gene_135 1.7259726  -18.55038 -15.851408       49
#> gene_325 1.2093850  -18.52829 -15.829324       84
#> gene_291 2.2015170  -18.30647 -15.607500       36
#> gene_488 3.0647537  -18.06703 -15.368055       22
#> gene_340 1.4521210  -17.81841 -15.119440       56
#> gene_213 1.0901441  -17.59101 -14.892037       85
#> gene_458 0.9508191  -17.54105 -14.842082       92
#> gene_295 1.7966484  -17.49376 -14.794790       42
#> gene_424 1.4472784  -17.47510 -14.776132       54
#> gene_51  1.6407593  -16.72013 -14.021162       43
#> gene_449 2.4810190  -16.22803 -13.529063       24
#> gene_99  1.0849556  -16.19768 -13.498709       71
#> gene_13  1.4534620  -14.76300 -12.064030       40
#> gene_10  1.3278710  -14.70646 -12.007492       44
#> gene_136 1.6923133  -13.85618 -11.157209       30
#> gene_216 1.7300714  -13.54051 -10.841541       28
#> gene_176 2.0046279  -13.30574 -10.606773       22
#> gene_156 1.4589040  -12.39352  -9.694550       30
#> gene_225 1.6076632  -11.42117  -8.722198       23
#> gene_201 1.2245458  -10.28770  -7.588726       28
#> gene_234 1.8291523  -10.12013  -7.421157       15

# 3. list a set of specified genes
set <- c("gene_497","gene_386", "gene_275")
show_result_haystack(res.haystack = res, gene = set)

#>               D_KL log.p.vals log.p.adj T.counts
#> gene_275 1.3452514 -22.338516 -19.63955       95
#> gene_497 1.5996456 -21.943885 -19.24491       86
#> gene_386 0.2531526  -1.512989   0.00000       30