show_result_haystack

Shows the results of the 'haystack' analysis in various ways, sorted by significance. Priority of params is genes > p.value.threshold > n.

show_result_haystack(
  res.haystack,
  n = NULL,
  p.value.threshold = NULL,
  gene = NULL
)

# S3 method for haystack
show_result_haystack(
  res.haystack,
  n = NULL,
  p.value.threshold = NULL,
  gene = NULL
)

Arguments

res.haystack

A 'haystack' result object.

n

If defined, the top "n" significant genes will be returned. Default: NA, which shows all results.

p.value.threshold

If defined, genes passing this p-value threshold will be returned.

gene

If defined, the results of this (these) gene(s) will be returned.

Value

A data.frame with 'haystack' results sorted by log.p.vals.

Details

The output is a data.frame with the following columns: * D_KL the calculated KL divergence. * log.p.vals log10 p.values calculated from randomization. * log.p.adj log10 p.values adjusted by Bonferroni correction.

Examples

# using the toy example of the singleCellHaystack package

# running haystack
res <- haystack(dat.tsne, dat.expression)
#> ### calling haystack_continuous_highD()...
#> ### Using package sparseMatrixStats to speed up statistics in sparse matrices.
#> ### Calculating row-wise mean and SD... 
#> ### Filtered 0 genes with zero variance...
#> ### Using 100 randomizations...
#> ### Using 100 genes to randomize...
#> Warning: The value of 'grid.points' appears to be very high (> No. of cells / 10). You can set the number of grid points using the 'grid.points' parameter.
#> ### scaling input data...
#> ### deciding grid points...
#> ### calculating Kullback-Leibler divergences...
#> ### performing randomizations...
#> ### estimating p-values...
#> ### picking model for mean D_KL...
#> ### using natural splines
#> ### best RMSD  : 0.089
#> ### best df    : 3
#> ### picking model for stdev D_KL...
#> ### using natural splines
#> ### best RMSD  : 0.018
#> ### best df    : 4
#> ### returning result...

# below are variations for showing the results in a table
# 1. list top 10 biased genes
show_result_haystack(res.haystack = res, n =10)
#>              D_KL log.p.vals log.p.adj
#> gene_62  2.109978  -36.89496 -34.19599
#> gene_61  2.095796  -35.75287 -33.05390
#> gene_79  2.208038  -34.46184 -31.76287
#> gene_242 1.875956  -34.45565 -31.75668
#> gene_99  2.017610  -33.06250 -30.36353
#> gene_444 1.963573  -32.83715 -30.13818
#> gene_339 1.868211  -32.79171 -30.09274
#> gene_351 1.868483  -32.08044 -29.38147
#> gene_497 1.871301  -31.57997 -28.88100
#> gene_71  2.505652  -31.38314 -28.68417
# 2. list genes with p value below a certain threshold
show_result_haystack(res.haystack = res, p.value.threshold=1e-10)
#>              D_KL log.p.vals  log.p.adj
#> gene_62  2.109978  -36.89496 -34.195985
#> gene_61  2.095796  -35.75287 -33.053904
#> gene_79  2.208038  -34.46184 -31.762874
#> gene_242 1.875956  -34.45565 -31.756680
#> gene_99  2.017610  -33.06250 -30.363526
#> gene_444 1.963573  -32.83715 -30.138178
#> gene_339 1.868211  -32.79171 -30.092744
#> gene_351 1.868483  -32.08044 -29.381473
#> gene_497 1.871301  -31.57997 -28.880996
#> gene_71  2.505652  -31.38314 -28.684166
#> gene_429 1.784322  -31.04067 -28.341704
#> gene_275 1.761348  -30.48882 -27.789849
#> gene_137 1.780374  -29.92616 -27.227192
#> gene_300 2.128255  -28.96860 -26.269629
#> gene_479 2.311933  -28.82584 -26.126866
#> gene_317 1.731105  -28.80559 -26.106623
#> gene_321 2.324771  -27.24310 -24.544133
#> gene_325 1.664610  -26.18602 -23.487048
#> gene_155 1.980354  -25.66228 -22.963308
#> gene_59  2.127830  -25.36601 -22.667043
#> gene_313 1.629849  -25.33257 -22.633603
#> gene_112 2.135417  -25.32436 -22.625386
#> gene_340 2.265802  -24.89204 -22.193071
#> gene_24  2.417165  -24.40300 -21.704035
#> gene_244 1.522756  -24.30693 -21.607956
#> gene_213 1.527215  -23.73415 -21.035176
#> gene_381 2.185395  -23.38588 -20.686909
#> gene_424 2.375324  -22.86083 -20.161855
#> gene_51  2.711843  -22.20605 -19.507085
#> gene_458 1.430809  -21.13419 -18.435222
#> gene_57  2.325346  -21.11508 -18.416106
#> gene_400 2.468799  -19.93427 -17.235296
#> gene_463 2.755731  -19.50091 -16.801936
#> gene_135 2.382106  -19.06350 -16.364528
#> gene_78  2.806871  -18.85727 -16.158300
#> gene_295 2.463360  -16.80082 -14.101846
#> gene_216 3.004364  -16.18816 -13.489192
#> gene_291 2.712180  -16.08182 -13.382853
#> gene_449 3.219535  -15.44642 -12.747448
#> gene_13  2.304398  -15.40043 -12.701464
#> gene_10  2.187587  -14.63827 -11.939304
#> gene_488 3.266444  -14.52657 -11.827604
#> gene_136 2.669582  -14.16005 -11.461082
#> gene_176 3.078777  -13.66620 -10.967233
#> gene_156 2.496897  -13.01641 -10.317436
#> gene_225 2.867361  -12.60903  -9.910056
#> gene_9   3.153705  -11.81336  -9.114391
#> gene_201 2.416615  -11.01261  -8.313642
#> gene_234 3.305842  -10.18107  -7.482105
# 3. list a set of specified genes
set <- c("gene_497","gene_386", "gene_275")
show_result_haystack(res.haystack = res, gene = set)
#>              D_KL  log.p.vals log.p.adj
#> gene_497 1.871301 -31.5799660 -28.88100
#> gene_275 1.761348 -30.4888187 -27.78985
#> gene_386 0.634804  -0.2093525   0.00000