show_result_haystack

Shows the results of the 'haystack' analysis in various ways, sorted by significance. Priority of params is genes > p.value.threshold > n.

show_result_haystack(
  res.haystack,
  n = NULL,
  p.value.threshold = NULL,
  gene = NULL
)

# S3 method for haystack
show_result_haystack(
  res.haystack,
  n = NULL,
  p.value.threshold = NULL,
  gene = NULL
)

Arguments

res.haystack

A 'haystack' result object.

n

If defined, the top "n" significant genes will be returned. Default: NA, which shows all results.

p.value.threshold

If defined, genes passing this p-value threshold will be returned.

gene

If defined, the results of this (these) gene(s) will be returned.

Value

A data.frame with 'haystack' results sorted by log.p.vals.

Details

The output is a data.frame with the following columns: * D_KL the calculated KL divergence. * log.p.vals log10 p.values calculated from randomization. * log.p.adj log10 p.values adjusted by Bonferroni correction.

Examples

# using the toy example of the singleCellHaystack package

# running haystack
res <- haystack(dat.tsne, dat.expression)
#> ### calling haystack_continuous_highD()...
#> ### Using package sparseMatrixStats to speed up statistics in sparse matrices.
#> ### Calculating row-wise mean and SD... 
#> ### Filtered 0 genes with zero variance...
#> ### Using 100 randomizations...
#> ### Using 100 genes to randomize...
#> Warning: The value of 'grid.points' appears to be very high (> No. of cells / 10). You can set the number of grid points using the 'grid.points' parameter.
#> ### scaling input data...
#> ### deciding grid points...
#> ### calculating Kullback-Leibler divergences...
#> ### performing randomizations...
#> ### estimating p-values...
#> ### picking model for mean D_KL...
#> ### using natural splines
#> ### best RMSD  : 0.089
#> ### best df    : 3
#> ### picking model for stdev D_KL...
#> ### using natural splines
#> ### best RMSD  : 0.015
#> ### best df    : 4
#> ### returning result...

# below are variations for showing the results in a table
# 1. list top 10 biased genes
show_result_haystack(res.haystack = res, n =10)
#>              D_KL log.p.vals log.p.adj
#> gene_242 1.900159  -36.34655 -33.64758
#> gene_79  2.237174  -34.97892 -32.27995
#> gene_61  2.068980  -34.81782 -32.11885
#> gene_62  1.989213  -33.88287 -31.18390
#> gene_444 1.988849  -33.20264 -30.50367
#> gene_275 1.785670  -31.88295 -29.18398
#> gene_351 1.840483  -31.81217 -29.11320
#> gene_497 1.862287  -31.46813 -28.76916
#> gene_99  1.934713  -30.92677 -28.22780
#> gene_429 1.745737  -30.89821 -28.19924
# 2. list genes with p value below a certain threshold
show_result_haystack(res.haystack = res, p.value.threshold=1e-10)
#>              D_KL log.p.vals  log.p.adj
#> gene_242 1.900159  -36.34655 -33.647584
#> gene_79  2.237174  -34.97892 -32.279953
#> gene_61  2.068980  -34.81782 -32.118854
#> gene_62  1.989213  -33.88287 -31.183905
#> gene_444 1.988849  -33.20264 -30.503667
#> gene_275 1.785670  -31.88295 -29.183979
#> gene_351 1.840483  -31.81217 -29.113204
#> gene_497 1.862287  -31.46813 -28.769163
#> gene_99  1.934713  -30.92677 -28.227805
#> gene_429 1.745737  -30.89821 -28.199244
#> gene_71  2.391586  -30.14138 -27.442407
#> gene_300 2.144801  -29.60178 -26.902812
#> gene_479 2.315855  -29.50101 -26.802044
#> gene_325 1.746532  -28.24227 -25.543296
#> gene_313 1.731401  -27.89502 -25.196054
#> gene_321 2.279258  -27.22289 -24.523918
#> gene_155 2.034497  -27.09020 -24.391226
#> gene_339 1.596495  -26.84914 -24.150174
#> gene_213 1.629577  -26.72606 -24.027091
#> gene_244 1.584252  -26.52728 -23.828311
#> gene_381 2.233813  -24.96956 -22.270592
#> gene_317 1.518662  -24.02622 -21.327254
#> gene_137 1.522466  -24.01957 -21.320604
#> gene_24  2.294962  -23.58092 -20.881955
#> gene_458 1.475453  -22.67433 -19.975358
#> gene_57  2.328396  -22.16051 -19.461544
#> gene_51  2.595802  -22.00362 -19.304652
#> gene_400 2.532468  -21.82195 -19.122975
#> gene_291 3.160579  -21.57637 -18.877403
#> gene_135 2.485737  -21.42512 -18.726154
#> gene_59  1.859318  -21.14511 -18.446140
#> gene_463 2.773410  -20.85622 -18.157246
#> gene_78  2.871959  -20.68588 -17.986915
#> gene_340 1.957126  -20.63622 -17.937250
#> gene_424 2.141243  -20.45695 -17.757982
#> gene_112 1.762119  -19.23574 -16.536767
#> gene_295 2.476866  -17.97710 -15.278129
#> gene_13  2.387906  -17.32226 -14.623286
#> gene_488 3.339279  -15.66320 -12.964230
#> gene_136 2.703013  -15.33182 -12.632854
#> gene_10  2.141630  -14.98755 -12.288576
#> gene_216 2.759649  -14.75644 -12.057471
#> gene_449 3.059416  -14.65668 -11.957711
#> gene_156 2.567836  -14.52540 -11.826433
#> gene_176 3.064945  -14.08711 -11.388139
#> gene_225 2.958600  -14.02095 -11.321975
#> gene_201 2.470913  -12.18803  -9.489063
#> gene_234 3.443085  -11.19281  -8.493844
#> gene_9   2.984528  -10.53198  -7.833011
# 3. list a set of specified genes
set <- c("gene_497","gene_386", "gene_275")
show_result_haystack(res.haystack = res, gene = set)
#>               D_KL  log.p.vals log.p.adj
#> gene_275 1.7856696 -31.8829492 -29.18398
#> gene_497 1.8622874 -31.4681333 -28.76916
#> gene_386 0.7448268  -0.8429364   0.00000