Shows the results of the 'haystack' analysis in various ways, sorted by significance. Priority of params is genes > p.value.threshold > n.

Shows the results of the 'haystack' analysis in various ways, sorted by significance. Priority of params is genes > p.value.threshold > n.

show_result_haystack(res.haystack, n = NA, p.value.threshold = NA, gene = NA)

Arguments

res.haystack
A 'haystack' result variable

n
If defined, the top "n" sigificant genes will be returned. Default: NA, which shows all results.

p.value.threshold
If defined, genes passing this p-value threshold will be returned.

gene
If defined, the results of this (these) gene(s) will be returned.

Value

A table with a sorted subset of the 'haystack' result according to input parameters.

Examples

# using the toy example of the singleCellHaystack package
# define a logical matrix with detection of each gene (rows) in each cell (columns)
dat.detection <- dat.expression > 1

# running haystack in default mode
res <- haystack(dat.tsne, detection=dat.detection, method = "2D")

#> ### calling haystack_2D()...

#> ### setting parameters...

#> ### calculating Kullback-Leibler divergences...

#> ### performing randomizations...

#> ### estimating p-values...

#> ### returning result...

# below are variations for showing the results in a table
# 1. list top 10 biased genes
show_result_haystack(res.haystack = res, n =10)

#>              D_KL log.p.vals log.p.adj T.counts
#> gene_497 1.599646  -22.57603 -19.87706       86
#> gene_79  1.929800  -21.56624 -18.86727       68
#> gene_62  1.519993  -21.42687 -18.72790       81
#> gene_351 1.408946  -21.22743 -18.52846       89
#> gene_339 1.414814  -21.20838 -18.50941       87
#> gene_61  1.470475  -20.89387 -18.19490       80
#> gene_317 1.386194  -20.85753 -18.15856       94
#> gene_137 1.349994  -20.74511 -18.04614       89
#> gene_242 1.360661  -20.43360 -17.73463       96
#> gene_275 1.345251  -20.43203 -17.73306       95

# 2. list genes with p value below a certain threshold
show_result_haystack(res.haystack = res, p.value.threshold=1e-10)

#>               D_KL log.p.vals  log.p.adj T.counts
#> gene_497 1.5996456  -22.57603 -19.877064       86
#> gene_79  1.9298000  -21.56624 -18.867272       68
#> gene_62  1.5199933  -21.42687 -18.727897       81
#> gene_351 1.4089458  -21.22743 -18.528459       89
#> gene_339 1.4148142  -21.20838 -18.509408       87
#> gene_61  1.4704748  -20.89387 -18.194900       80
#> gene_317 1.3861936  -20.85753 -18.158555       94
#> gene_137 1.3499940  -20.74511 -18.046136       89
#> gene_242 1.3606607  -20.43360 -17.734630       96
#> gene_275 1.3452514  -20.43203 -17.733063       95
#> gene_444 1.4358815  -19.73397 -17.035004       75
#> gene_244 1.2380108  -19.54806 -16.849087       95
#> gene_325 1.2093850  -19.21995 -16.520977       84
#> gene_24  1.8574968  -19.21814 -16.519172       59
#> gene_313 1.2945440  -19.13959 -16.440622       78
#> gene_71  1.8688220  -18.73361 -16.034643       56
#> gene_479 1.6972565  -18.63387 -15.934904       61
#> gene_300 1.6001648  -18.61041 -15.911443       64
#> gene_155 1.5347601  -18.37397 -15.674996       65
#> gene_213 1.0901441  -18.19559 -15.496620       85
#> gene_429 1.0719883  -18.14001 -15.441041       94
#> gene_59  1.4605859  -18.05834 -15.359372       66
#> gene_57  1.8012672  -17.99210 -15.293128       54
#> gene_321 1.5354565  -17.97202 -15.273046       63
#> gene_381 1.6793541  -17.96130 -15.262335       58
#> gene_78  2.3302373  -17.63274 -14.933769       38
#> gene_463 2.1619665  -17.50940 -14.810431       41
#> gene_400 1.8579716  -17.27257 -14.573596       48
#> gene_458 0.9508191  -17.01695 -14.317983       92
#> gene_112 1.4955762  -16.94732 -14.248353       59
#> gene_135 1.7259726  -16.69053 -13.991556       49
#> gene_291 2.2015170  -16.54678 -13.847812       36
#> gene_340 1.4521210  -16.13515 -13.436179       56
#> gene_99  1.0849556  -16.01521 -13.316240       71
#> gene_295 1.7966484  -15.80601 -13.107037       42
#> gene_424 1.4472784  -15.77061 -13.071643       54
#> gene_488 3.0647537  -15.54690 -12.847934       22
#> gene_51  1.6407593  -15.10192 -12.402946       43
#> gene_449 2.4810190  -14.16364 -11.464666       24
#> gene_13  1.4534620  -13.36608 -10.667108       40
#> gene_10  1.3278710  -13.28632 -10.587346       44
#> gene_136 1.6923133  -12.42545  -9.726482       30
#> gene_216 1.7300714  -12.06586  -9.366890       28
#> gene_176 2.0046279  -11.47633  -8.777359       22
#> gene_156 1.4589040  -11.12358  -8.424612       30

# 3. list a set of specified genes
set <- c("gene_497","gene_386", "gene_275")
show_result_haystack(res.haystack = res, gene = set)

#>               D_KL log.p.vals log.p.adj T.counts
#> gene_497 1.5996456 -22.576034 -19.87706       86
#> gene_275 1.3452514 -20.432033 -17.73306       95
#> gene_386 0.2531526  -1.409232   0.00000       30