show_result_haystack

Shows the results of the 'haystack' analysis in various ways, sorted by significance. Priority of params is genes > p.value.threshold > n.

show_result_haystack(res.haystack, n = NA, p.value.threshold = NA,
  gene = NA)

Arguments

res.haystack
A 'haystack' result variable

n
If defined, the top "n" sigificant genes will be returned. Default: NA, which shows all results.

p.value.threshold
If defined, genes passing this p-value threshold will be returned.

gene
If defined, the results of this (these) gene(s) will be returned.

Value

A table with a sorted subset of the 'haystack' result according to input parameters.

Examples

# using the toy example of the singleCellHaystack package
# define a logical matrix with detection of each gene (rows) in each cell (columns)
dat.detection <- dat.expression > 1

# running haystack in default mode
res <- haystack(dat.tsne, detection=dat.detection, method = "2D")

#> ### calling haystack_2D()...

#> ### setting parameters...

#> ### calculating Kullback-Leibler divergences...

#> ### starting randomizations...

#> ### ... 10 values out of 58 done

#> ### ... 20 values out of 58 done

#> ### ... 30 values out of 58 done

#> ### ... 40 values out of 58 done

#> ### ... 50 values out of 58 done

#> ### estimating p-values...

#> ### returning result...

# below are variations for showing the results in a table
# 1. list top 10 biased genes
show_result_haystack(res.haystack = res, n =10)

#>              D_KL log.p.vals T.counts
#> gene_497 1.599646  -22.31043       86
#> gene_62  1.519993  -20.90104       81
#> gene_351 1.408946  -20.84695       89
#> gene_339 1.414814  -20.80586       87
#> gene_79  1.929800  -20.45369       68
#> gene_317 1.386194  -20.40589       94
#> gene_137 1.349994  -20.30840       89
#> gene_61  1.470475  -20.27907       80
#> gene_275 1.345251  -19.90779       95
#> gene_242 1.360661  -19.88499       96

# 2. list genes with p value below a certain threshold
show_result_haystack(res.haystack = res, p.value.threshold=1e-10)

#>               D_KL log.p.vals T.counts
#> gene_497 1.5996456  -22.31043       86
#> gene_62  1.5199933  -20.90104       81
#> gene_351 1.4089458  -20.84695       89
#> gene_339 1.4148142  -20.80586       87
#> gene_79  1.9298000  -20.45369       68
#> gene_317 1.3861936  -20.40589       94
#> gene_137 1.3499940  -20.30840       89
#> gene_61  1.4704748  -20.27907       80
#> gene_275 1.3452514  -19.90779       95
#> gene_242 1.3606607  -19.88499       96
#> gene_244 1.2380108  -18.91327       95
#> gene_444 1.4358815  -18.80945       75
#> gene_325 1.2093850  -18.54440       84
#> gene_313 1.2945440  -18.27911       78
#> gene_24  1.8574968  -17.51135       59
#> gene_213 1.0901441  -17.42695       85
#> gene_429 1.0719883  -17.35265       94
#> gene_300 1.6001648  -17.09222       64
#> gene_479 1.6972565  -16.98287       61
#> gene_155 1.5347601  -16.88542       65
#> gene_71  1.8688220  -16.87939       56
#> gene_59  1.4605859  -16.59295       66
#> gene_321 1.5354565  -16.36856       63
#> gene_381 1.6793541  -16.14979       58
#> gene_458 0.9508191  -16.12652       92
#> gene_57  1.8012672  -16.03445       54
#> gene_78  2.3302373  -15.37371       38
#> gene_463 2.1619665  -15.25658       41
#> gene_112 1.4955762  -15.12303       59
#> gene_400 1.8579716  -15.12158       48
#> gene_99  1.0849556  -14.60397       71
#> gene_135 1.7259726  -14.54798       49
#> gene_488 3.0647537  -14.38967       22
#> gene_291 2.2015170  -14.28109       36
#> gene_340 1.4521210  -14.16820       56
#> gene_424 1.4472784  -13.72857       54
#> gene_295 1.7966484  -13.53518       42
#> gene_51  1.6407593  -12.82867       43
#> gene_449 2.4810190  -12.53533       24
#> gene_13  1.4534620  -11.04548       40
#> gene_10  1.3278710  -10.99048       44
#> gene_136 1.6923133  -10.20170       30

# 3. list a set of specified genes
set <- c("gene_497","gene_386", "gene_275")
show_result_haystack(res.haystack = res, gene = set)

#>               D_KL log.p.vals T.counts
#> gene_497 1.5996456  -22.31043       86
#> gene_275 1.3452514  -19.90779       95
#> gene_386 0.2531526    0.00000       30